projet_web/src/requirements.txt
pymysql
mysqlclient
mysql-connector-python

gunicorn
Flask
flask-sqlalchemy

flask-wtf
wtforms-html5
wtf-peewee